add FF and bridge flows
