+++ packages/web/components/bridge/immersive/index.ts
@@ -1,0 +1,1 @@
+export * from "./provider";
